Adicionei as funções dos gets e set das rules na gui e já faz o load das rules para a GUI.

+++ src/antiSpamFilter/Rule.java
@@ -3,9 +3,9 @@ package antiSpamFilter;
 public class Rule {
 
 	private String name;
-	private int weight;
+	private double weight;
 	
-	public Rule(String name, int weight) {
+	public Rule(String name, double weight) {
 		this.name=name;
 		this.weight=weight;
 		
@@ -20,11 +20,11 @@ public class Rule {
 		this.name = name;
 	}
 
-	public int getWeight() {
+	public double getWeight() {
 		return weight;
 	}
 
-	public void setWeight(int weight) {
+	public void setWeight(double weight) {
 		this.weight = weight;
 	}
